MiddlewareEnabledSequentialAsyncMessageRouter and MessageEnvelope
Lets you hook into the route pipe in similar way as Owin.

Use MessageEnvelope in internal routes. The handlers and interfaces
etc. in your app are still using plain messages. But internally
a MessageEnvelope is being passed. This so that you e.g. can construct
middlewares that attaches state etc to the envelope. This state can then
be used when resolving MessageHandler class instances, for e.g.
extracting a Autofac scope from the envelope.